refactor(auth): refactor email verification process to use subdomain for CMS member password reset
